DiagrammeROutput {DiagrammeR}	R Documentation

Widget output function for use in Shiny

Description

Widget output function for use in Shiny

Usage

DiagrammeROutput(outputId, width = "100%", height = "auto")

Arguments

outputId
output variable to read from

width
a valid CSS unit for the width or a number, which will be coerced to a string and have px appended.

height
a valid CSS unit for the height or a number, which will be coerced to a string and have px appended.

[Package DiagrammeR version 1.0.1 Index]