add_global_graph_attrs {DiagrammeR}	R Documentation

Add one or more global graph attributes

Description

Add global attributes of a specific type (either graph_attrs, node_attrs, or edge_attrs for a graph object of class dgr_graph).

Usage

add_global_graph_attrs(graph, attr, value, attr_type)

Arguments

graph
a graph object of class dgr_graph.

attr
the name of the attribute to set for the type of global attribute specified.

value
the value to be set for the chosen attribute specified in the attr_for_type argument.

attr_type
the specific type of global graph attribute to set. The type is specified with graph, node, or edge.

Value

a graph object of class dgr_graph.

Examples

# Create a new graph with no
# global graph attributes and
# add a global graph attribute
graph <-
  create_graph(
    attr_theme = NULL) %>%
  add_global_graph_attrs(
    attr = "overlap",
    value = "true",
    attr_type = "graph")

# Verify that the attribute
# addition has been made
graph %>%
  get_global_graph_attr_info()

# Add another attribute with
# `add_global_graph_attrs()`
graph <-
  graph %>%
  add_global_graph_attrs(
    attr = "penwidth",
    value = 12,
    attr_type = "node")

# Verify that the attribute
# addition has been made
graph %>%
  get_global_graph_attr_info()

# When adding an attribute where
# `attr` and `attr_type` already
# exists, the value provided will
# serve as an update
graph %>%
  add_global_graph_attrs(
    attr = "penwidth",
    value = 15,
    attr_type = "node") %>%
  get_global_graph_attr_info()

[Package DiagrammeR version 1.0.1 Index]