add_islands_graph {DiagrammeR}	R Documentation

Create a random islands graph with edges between the islands

Description

To an existing graph object, add several Erdos-Renyi random graphs (the islands) using a common set of parameters, connected together by a fixed number of edges.

Usage

add_islands_graph(graph, n_islands, island_size, p, edges_between,
  type = NULL, label = TRUE, rel = NULL, node_aes = NULL,
  edge_aes = NULL, node_data = NULL, edge_data = NULL, set_seed = NULL)

Arguments

graph
a graph object of class dgr_graph.

n_islands
the number of islands in the generated graph.

island_size
the size of the islands in the generated graph.

p
the probability of there being edges between the islands.

edges_between
The number of edges between islands.

type
an optional string that describes the entity type for all the nodes to be added.

label
a boolean value where setting to TRUE ascribes node IDs to the label and FALSE yields a blank label.

rel
an optional string for providing a relationship label to all edges to be added.

node_aes
an optional list of named vectors comprising node aesthetic attributes. The helper function node_aes() is strongly recommended for use here as it contains arguments for each of the accepted node aesthetic attributes (e.g., shape, style, color, fillcolor).

edge_aes
an optional list of named vectors comprising edge aesthetic attributes. The helper function edge_aes() is strongly recommended for use here as it contains arguments for each of the accepted edge aesthetic attributes (e.g., shape, style, penwidth, color).

node_data
an optional list of named vectors comprising node data attributes. The helper function node_data() is strongly recommended for use here as it helps bind data specifically to the created nodes.

edge_data
an optional list of named vectors comprising edge data attributes. The helper function edge_data() is strongly recommended for use here as it helps bind data specifically to the created edges.

set_seed
supplying a value sets a random seed of the Mersenne-Twister implementation.

Examples

# Create a graph of islands
islands_graph <-
  create_graph() %>%
  add_islands_graph(
    n_islands = 4,
    island_size = 10,
    p = 0.5,
    edges_between = 1,
    set_seed = 23)

# Get a count of nodes
islands_graph %>%
  count_nodes()

# Get a count of edges
islands_graph %>%
  count_edges()

[Package DiagrammeR version 1.0.1 Index]