add_n_nodes {DiagrammeR}	R Documentation

Add one or several unconnected nodes to the graph

Description

Add n new nodes to a graph object of class dgr_graph. Optionally, set node type values for the new nodes.

Usage

add_n_nodes(graph, n, type = NULL, label = NULL, node_aes = NULL,
  node_data = NULL)

Arguments

graph
a graph object of class dgr_graph.

n
the number of new nodes to add to the graph.

type
an optional character vector that provides group identifiers for the nodes to be added.

label
an optional character object that describes the nodes to be added.

node_aes
an optional list of named vectors comprising node aesthetic attributes. The helper function node_aes() is strongly recommended for use here as it contains arguments for each of the accepted node aesthetic attributes (e.g., shape, style, color, fillcolor).

node_data
an optional list of named vectors comprising node data attributes. The helper function node_data() is strongly recommended for use here as it helps bind data specifically to the created nodes.

Value

a graph object of class dgr_graph.

Examples

# Create an empty graph and
# add 5 nodes; these nodes
# will be assigned ID values
# from `1` to `5`
graph <-
  create_graph() %>%
  add_n_nodes(n = 5)

# Get the graph's node IDs
graph %>%
  get_node_ids()

[Package DiagrammeR version 1.0.1 Index]