create_edge_df {DiagrammeR}	R Documentation

Create an edge data frame

Description

Combine several vectors for edges and their attributes into a data frame, which can be combined with other similarly-generated data frames, or, added to a graph object. An edge data frame, or edf, has at least the following columns:

- id (of type integer)

- from (of type integer)

- to (of type integer)

- rel (of type character)

An arbitrary number of additional columns containing aesthetic or data attributes can be part of the edf, so long as they follow the aforementioned columns.

Usage

create_edge_df(from, to, rel = NULL, ...)

Arguments

from
a vector of node ID values from which edges are outbound. The vector length must equal that of the to vector.

to
a vector of node ID values to which edges are incoming. The vector length must equal that of the from vector.

rel
an optional rel label for each edge.

...
one or more vectors for associated edge attributes.

Value

an edge data frame (edf).

Examples

# Create a simple edge data frame (edf) and
# view the results
edf <-
  create_edge_df(
    from = c(1, 2, 3),
    to = c(4, 3, 1),
    rel = "a")

# Display the edge data frame
edf

# Create an edf with additional edge
# attributes (where their classes will
# be inferred from the input vectors)
edf <-
  create_edge_df(
    from = c(1, 2, 3),
    to = c(4, 3, 1),
    rel = "a",
    length = c(50, 100, 250),
    color = "green",
    width = c(1, 5, 2))

# Display the edge data frame
edf

[Package DiagrammeR version 1.0.1 Index]