create_node_df {DiagrammeR}	R Documentation

Create a node data frame

Description

Combine several vectors for nodes and their attributes into a data frame, which can be combined with other similarly-generated data frames, or, added to a graph object. A node data frame, or ndf, has at least the following columns:

- id (of type integer)

- type (of type character)

- label (of type character)

An arbitrary number of additional columns containing aesthetic or data attributes can be part of the ndf, so long as they follow the aforementioned columns.

Usage

create_node_df(n, type = NULL, label = NULL, ...)

Arguments

n
the total number of nodes to include in the node data frame.

type
an optional type for each node.

label
an optional label for each node.

...
one or more vectors for associated node attributes.

Value

a node data frame (ndf).

Examples

# Create a node data frame (ndf) where the labels
# are equivalent to the node ID values (this is not
# recommended); the `label` and `type` node
# attributes will always be a `character` class
# whereas `id` will always be an `integer`
node_df <-
  create_node_df(
    n = 4,
    type = c("a", "a", "b", "b"),
    label = TRUE)

# Display the node data frame
node_df

# Create an ndf with distinct labels and
# additional node attributes (where their classes
# will be inferred from the input vectors)
node_df <-
  create_node_df(
    n = 4,
    type = "a",
    label = c(2384, 3942, 8362, 2194),
    style = "filled",
    color = "aqua",
    shape = c("circle", "circle",
              "rectangle", "rectangle"),
    value = c(3.5, 2.6, 9.4, 2.7))

# Display the node data frame
node_df

[Package DiagrammeR version 1.0.1 Index]