delete_global_graph_attrs {DiagrammeR}	R Documentation

Delete one of the global graph attributes stored within a graph object

Description

Delete one of the global attributes stored within a graph object of class dgr_graph).

Usage

delete_global_graph_attrs(graph, attr = NULL, attr_type = NULL)

Arguments

graph
a graph object of class dgr_graph.

attr
the name of the attribute to delete for the type of global attribute specified.

attr_type
the specific type of global graph attribute to delete. The type is specified with graph, node, or edge.

Value

a graph object of class dgr_graph.

Examples

# Create a new graph and add
# some extra global graph attrs
graph <-
  create_graph() %>%
  add_global_graph_attrs(
    attr = "overlap",
    value = "true",
    attr_type = "graph") %>%
  add_global_graph_attrs(
    attr = "penwidth",
    value = 3,
    attr_type = "node") %>%
  add_global_graph_attrs(
    attr = "penwidth",
    value = 3,
    attr_type = "edge")

# Inspect the graph's global
# attributes
graph %>%
  get_global_graph_attr_info()

# Delete the `penwidth` attribute
# for the graph's nodes using the
# `delete_global_graph_attrs()` fcn
graph <-
  graph %>%
  delete_global_graph_attrs(
    attr = "penwidth",
    attr_type = "node")

# View the remaining set of global
# attributes for the graph
graph %>%
  get_global_graph_attr_info()

[Package DiagrammeR version 1.0.1 Index]