edge_list_2 {DiagrammeR}	R Documentation

Edge list - Version 2.

Description

A simple, 5-column data frame that can be used to generate graph edges.

Usage

edge_list_2

Format

A data frame with 19 rows and 5 variables:

from

integer values that state the node ID values where an edge starts

to

integer values that state the node ID values where an edge terminates

rel

a grouping variable of either a, b, or c

value_1

a randomized set of numeric values between 0 and 10

value_2

a randomized set of numeric values between 0 and 10

[Package DiagrammeR version 1.0.1 Index]