from_adj_matrix {DiagrammeR}	R Documentation

Create a graph using an adjacency matrix

Description

Using an adjacency matrix object, generate a graph of class dgr_graph.

Usage

from_adj_matrix(x, mode = "undirected", weighted = FALSE, use_diag = TRUE,
  graph_name = NULL, write_backups = FALSE)

Arguments

x
a square matrix object serving as the adjacency matrix.

mode
the method in which to interpret the input adjacency matrix. Options include: undirected, directed, upper, lower, max, min, and plus.

weighted
whether to create a weighted graph from the adjacency matrix.

use_diag
whether to use the diagonal of the adjacency matrix in calculations. If TRUE then the diagonal values will be included as is. If FALSE then the diagonal values will be replaced with zero values before inclusion in any calculations.

graph_name
an optional string for labeling the graph object.

write_backups
an option to write incremental backups of changing graph states to disk. If TRUE, a subdirectory of the working directory will be used to store RDS files. The default value is FALSE so one has to opt in to use this functionality.

Value

a graph object of class dgr_graph.

Examples

# Create an adjacency matrix
adj_matrix <-
  sample(0:1, 100,
         replace = TRUE,
         prob = c(0.9,0.1)) %>%
  matrix(nc = 10)

# Create a graph from the adjacency matrix
graph <-
  from_adj_matrix(adj_matrix)

[Package DiagrammeR version 1.0.1 Index]