get_adhesion {DiagrammeR}	R Documentation

Get graph adhesion

Description

Get the adhesion of a graph, which is the minimum number of edges needed to remove to obtain a graph which is not strongly connected. This is the same as the edge connectivity of the graph.

Usage

get_adhesion(graph)

Arguments

graph
a graph object of class dgr_graph.

Value

a single numeric value representing the minimum number of edges to remove.

Examples

# Create a cycle graph
graph <-
  create_graph() %>%
  add_cycle(n = 5)

# Determine the graph's adhesion
graph %>%
  get_adhesion()

# Create a full graph and then
# get the adhesion for that
create_graph() %>%
  add_full_graph(n = 8) %>%
  get_adhesion()

[Package DiagrammeR version 1.0.1 Index]