get_alpha_centrality {DiagrammeR}	R Documentation

Get the alpha centrality for all nodes

Description

Get the alpha centrality values for all nodes in the graph.

Usage

get_alpha_centrality(graph, alpha = 1, exo = 1, weights_attr = NULL,
  tol = 1e-07)

Arguments

graph
a graph object of class dgr_graph.

alpha
the parameter that specifies the relative importance of endogenous versus exogenous factors in the determination of centrality.

exo
the exogenous factors, in most cases this is either a constant (which applies the same factor to every node), or a vector giving the factor for every node.

weights_attr
an optional name of the edge attribute to use in the adjacency matrix. If NULL then, if it exists, the weight edge attribute of the graph will be used. Failing that, the standard adjacency matrix will be used in calculations.

tol
the tolerance for near-singularities during matrix inversion. Default value is set to 1e-7.

Value

a data frame with alpha centrality scores for each of the nodes.

Examples

# Create a random graph using the
# `add_gnm_graph()` function
graph <-
  create_graph() %>%
  add_gnm_graph(
    n = 10,
    m = 12,
    set_seed = 23)

# Get the alpha centrality scores
# for all nodes
graph %>%
  get_alpha_centrality()

# Add the alpha centrality
# scores to the graph as a node
# attribute
graph <-
  graph %>%
  join_node_attrs(
    df = get_alpha_centrality(.))

# Display the graph's node
# data frame
graph %>%
  get_node_df()

[Package DiagrammeR version 1.0.1 Index]