get_authority_centrality {DiagrammeR}	R Documentation

Get the authority scores for all nodes

Description

Get the Kleinberg authority centrality scores for all nodes in the graph.

Usage

get_authority_centrality(graph, weights_attr = NULL)

Arguments

graph
a graph object of class dgr_graph.

weights_attr
an optional name of the edge attribute to use in the adjacency matrix. If NULL then, if it exists, the weight edge attribute of the graph will be used.

Value

a data frame with authority scores for each of the nodes.

Examples

# Create a random graph using the
# `add_gnm_graph()` function
graph <-
  create_graph() %>%
  add_gnm_graph(
    n = 10,
    m = 15,
    set_seed = 23)

# Get the authority centrality scores
# for all nodes in the graph
graph %>%
  get_authority_centrality()

# Add the authority centrality
# scores to the graph as a node
# attribute
graph <-
  graph %>%
  join_node_attrs(
    df = get_authority_centrality(.))

# Display the graph's node data frame
graph %>%
  get_node_df()

[Package DiagrammeR version 1.0.1 Index]