get_closeness {DiagrammeR}	R Documentation

Get closeness centrality values

Description

Get the closeness centrality values for all nodes in a graph.

Usage

get_closeness(graph, direction = "all")

Arguments

graph
a graph object of class dgr_graph.

direction
using all (the default), the search will ignore edge direction while traversing through the graph. With out, measurements of paths will be from a node whereas with in, measurements of paths will be to a node.

Value

a data frame with closeness values for each of the nodes.

Examples

# Create a random graph using the
# `add_gnm_graph()` function
graph <-
  create_graph() %>%
  add_gnm_graph(
    n = 10,
    m = 12,
    set_seed = 23)

# Get closeness values for all nodes
# in the graph
graph %>%
  get_closeness()

# Add the closeness values to
# the graph as a node attribute
graph <-
  graph %>%
  join_node_attrs(
    df = get_closeness(.))

# Display the graph's node data frame
graph %>%
  get_node_df()

[Package DiagrammeR version 1.0.1 Index]