get_cmty_walktrap {DiagrammeR}	R Documentation

Get community membership using the Walktrap method

Description

With the Walktrap community finding algorithm, obtain the group membership values for each of the nodes in the graph.

Usage

get_cmty_walktrap(graph, steps = 4)

Arguments

graph
a graph object of class dgr_graph.

steps
the number of steps to take for each of the random walks.

Value

a data frame with group membership assignments for each of the nodes.

Examples

# Create a random graph using the
# `add_gnm_graph()` function
graph <-
  create_graph(
    directed = FALSE) %>%
  add_gnm_graph(
    n = 10,
    m = 15,
    set_seed = 23)

# Get the group membership
# values for all nodes in the
# graph through the Walktrap
# community finding algorithm
graph %>%
  get_cmty_walktrap()

# Add the group membership
# values to the graph as a
# node attribute
graph <-
  graph %>%
  join_node_attrs(
    df = get_cmty_walktrap(.))

# Display the graph's
# node data frame
graph %>%
  get_node_df()

[Package DiagrammeR version 1.0.1 Index]