get_common_nbrs {DiagrammeR}	R Documentation

Get all common neighbors between two or more nodes

Description

With two or more nodes, get the set of common neighboring nodes.

Usage

get_common_nbrs(graph, nodes)

Arguments

graph
a graph object of class dgr_graph.

nodes
a vector of node ID values of length at least 2.

Value

a vector of node ID values.

Examples

# Create a directed graph with 5 nodes
graph <-
  create_graph() %>%
  add_path(n = 5)

# Find all common neighbor nodes
# for nodes `1` and `2` (there are no
# common neighbors amongst them)
graph %>%
  get_common_nbrs(
    nodes = c(1, 2))

# Find all common neighbor nodes for
# nodes `1` and `3`
graph %>%
  get_common_nbrs(
    nodes = c(1, 3))

[Package DiagrammeR version 1.0.1 Index]