get_degree_total {DiagrammeR}	R Documentation

Get total degree values for all nodes

Description

Get the total degree values for all nodes in a graph.

Usage

get_degree_total(graph, normalized = FALSE)

Arguments

graph
a graph object of class dgr_graph.

normalized
set as FALSE (the default), the total degree will be provided for each of the nodes (as a count of edges to and from each node). When set as TRUE, then the result for each node will be divided by the total number of nodes in the graph minus 1.

Value

a data frame with total degree values for each of the nodes.

Examples

# Create a random graph using the
# `add_gnm_graph()` function
graph <-
  create_graph(
    directed = FALSE) %>%
  add_gnm_graph(
    n = 10,
    m = 15,
    set_seed = 23)

# Get the total degree values
# for all nodes in the graph
graph %>%
  get_degree_total()

# Add the total degree values
# to the graph as a node
# attribute
graph <-
  graph %>%
  join_node_attrs(
    df = get_degree_total(.))

# Display the graph's
# node data frame
graph %>%
  get_node_df()

[Package DiagrammeR version 1.0.1 Index]