get_eccentricity {DiagrammeR}	R Documentation

Get node eccentricities

Description

Get a data frame with node eccentricity values.

Usage

get_eccentricity(graph, mode = "out")

Arguments

graph
a graph object of class dgr_graph.

mode
the mode with which the shortest paths to or from the given vertices should be calculated for directed graphs. If out (the default) then the shortest paths from the node, if in then only shortest paths to each node are considered. If all is used, then the corresponding undirected graph will be used and edge directions will be ignored. For undirected graphs, this argument is ignored.

Value

a data frame containing eccentricity values by node ID value.

Examples

# Create a random graph using the
# `add_gnm_graph()` function
graph <-
  create_graph(
    directed = FALSE) %>%
  add_gnm_graph(
    n = 10,
    m = 15,
    set_seed = 23)

# Get the eccentricity values for
# all nodes in the graph
graph %>%
  get_eccentricity()

[Package DiagrammeR version 1.0.1 Index]