get_eigen_centrality {DiagrammeR}	R Documentation

Get the eigen centrality for all nodes

Description

Get the eigen centrality values for all nodes in the graph.

Usage

get_eigen_centrality(graph, weights_attr = NULL)

Arguments

graph
a graph object of class dgr_graph.

weights_attr
an optional name of the edge attribute to use in the adjacency matrix. If NULL then, if it exists, the weight edge attribute of the graph will be used. If NA then no edge weights will be used.

Value

a data frame with eigen centrality scores for each of the nodes.

Examples

# Create a random graph using the
# `add_gnm_graph()` function
graph <-
  create_graph(
    directed = FALSE) %>%
  add_gnm_graph(
    n = 10, m = 15,
    set_seed = 23)

# Get the eigen centrality scores
# for nodes in the graph
graph %>%
  get_eigen_centrality()

[Package DiagrammeR version 1.0.1 Index]