get_global_graph_attr_info {DiagrammeR}	R Documentation

Get global graph attributes

Description

Get the available global attributes for a graph object of class dgr_graph.

Usage

get_global_graph_attr_info(graph)

Arguments

graph
a graph object of class dgr_graph.

Value

a data frame containing global attributes for the graph.

Examples

# Create a new, empty graph
graph <- create_graph()

# View the graph's set of
# global attributes
graph %>%
  get_global_graph_attr_info()

[Package DiagrammeR version 1.0.1 Index]