get_graph_name {DiagrammeR}	R Documentation

Get graph name

Description

Get the name of a graph object of class dgr_graph.

Usage

get_graph_name(graph)

Arguments

graph
a graph object of class dgr_graph.

Value

a single-length character vector with the assigned graph name. If a graph name has not been set, NA is returned.

Examples

# Create an empty graph
graph <- create_graph()

# Provide the new graph with a name
graph <-
  set_graph_name(
    graph,
    name = "the_name")

# Get the graph's name
graph %>%
  get_graph_name()

[Package DiagrammeR version 1.0.1 Index]