get_graph_series_info {DiagrammeR}	R Documentation

Get information on a graph series

Description

Obtain a data frame with information on the graphs within a graph series.

Usage

get_graph_series_info(graph_series)

Arguments

graph_series
a graph series object of type dgr_graph_1D.

Value

a data frame containing information on the graphs within the supplied graph series.

Examples

# Create three graphs
graph_1 <-
  create_graph() %>%
  add_path(n = 4)

graph_2 <-
  create_graph() %>%
  add_cycle(n = 5)

graph_3 <-
  create_graph() %>%
  add_star(n = 6)

# Create an empty graph series
# and add the graphs
series <-
  create_graph_series() %>%
  add_graph_to_graph_series(
    graph = graph_1) %>%
  add_graph_to_graph_series(
    graph = graph_2) %>%
  add_graph_to_graph_series(
    graph = graph_3)

# Get information on the graphs in the series
series %>%
  get_graph_series_info()

[Package DiagrammeR version 1.0.1 Index]