get_node_info {DiagrammeR}	R Documentation

Get detailed information on nodes

Description

Obtain a data frame with detailed information on nodes and their interrelationships within the graph.

Usage

get_node_info(graph)

Arguments

graph
a graph object of class dgr_graph.

Value

a data frame containing information specific to each node within the graph.

Examples

# Create a simple graph
graph <-
  create_graph() %>%
  add_gnm_graph(
    n = 5, m = 10,
    set_seed = 23)

# Get information on the
# graph's nodes
graph %>%
  get_node_info()

[Package DiagrammeR version 1.0.1 Index]