get_radiality {DiagrammeR}	R Documentation

Get radiality centrality scores

Description

Get the radiality centrality for all nodes in a graph. These scores describe the ease to which nodes can reach other nodes.

Usage

get_radiality(graph, direction = "all")

Arguments

graph
a graph object of class dgr_graph.

direction
using all (the default), the search will ignore edge direction while traversing through the graph. With out, measurements of paths will be from a node whereas with in, measurements of paths will be to a node.

Value

a data frame with radiality centrality scores for each of the nodes.

Examples

# Create a random graph using the
# `add_gnm_graph()` function
graph <-
  create_graph() %>%
  add_gnm_graph(
    n = 10,
    m = 15,
    set_seed = 23)

# Get the radiality scores for nodes in the graph
graph %>%
  get_radiality()

# Add the radiality values
# to the graph as a node
# attribute
graph <-
  graph %>%
  join_node_attrs(
    df = get_radiality(.))

# Display the graph's node data frame
graph %>%
  get_node_df()

[Package DiagrammeR version 1.0.1 Index]