grViz {DiagrammeR}	R Documentation

R + viz.js

Description

Make diagrams in R using viz.js with infrastructure provided by htmlwidgets.

Usage

grViz(diagram = "", engine = "dot", allow_subst = TRUE, options = NULL,
  width = NULL, height = NULL)

Arguments

diagram
spec for a diagram as either text, filename string, or file connection.

engine
string for the Graphviz layout engine; can be dot (default), neato, circo, or twopi. For more information see https://github.com/mdaines/viz.js#usage.

allow_subst
a boolean that enables/disables substitution functionality.

options
parameters supplied to the htmlwidgets framework.

width
an optional parameter for specifying the width of the resulting graphic in pixels.

height
an optional parameter for specifying the height of the resulting graphic in pixels.

Value

An object of class htmlwidget that will intelligently print itself into HTML in a variety of contexts including the R console, within R Markdown documents, and within Shiny output bindings.

[Package DiagrammeR version 1.0.1 Index]