import_graph {DiagrammeR}	R Documentation

Import a graph from various graph formats

Description

Import a variety of graphs from different graph formats and create a graph object.

Usage

import_graph(graph_file, file_type = NULL, edges_extra_attr_names = NULL,
  edges_extra_attr_coltypes = NULL)

Arguments

graph_file
a connection to a graph file. When provided as a path to a file, it will read the file from disk. Files starting with http://, https://, ftp://, or ftps:// will be automatically downloaded.

file_type
the type of file to be imported. Options are: gml (GML), sif (SIF), edges (a .edges file), and mtx (MatrixMarket format). If not supplied, the type of graph file will be inferred by its file extension.

edges_extra_attr_names
for edges files, a vector of attribute names beyond the from and to data columns can be provided in the order they appear in the input data file.

edges_extra_attr_coltypes
for edges files, this is a string of column types for any attribute columns provided for edges_extra_attr_names. This string representation is where each character represents each of the extra columns of data and the mappings are: c -> character, i -> integer, n -> number, d -> double, l -> logical, D -> date, T -> date time, t -> time, ? -> guess, or _/-, which skips the column.

Value

a graph object of class dgr_graph.

Examples

## Not run: 
# Import a GML graph file
gml_graph <-
  import_graph(
    system.file(
      "extdata/karate.gml",
      package = "DiagrammeR"))

# Get a count of the graph's nodes
gml_graph %>%
  count_nodes()

# Get a count of the graph's edges
gml_graph %>%
  count_edges()

## End(Not run)

[Package DiagrammeR version 1.0.1 Index]