is_graph_directed {DiagrammeR}	R Documentation

Is the graph a directed graph?

Description

Determines whether a graph is set to be directed or not and returns a logical value to that effect.

Usage

is_graph_directed(graph)

Arguments

graph
a graph object of class dgr_graph.

Value

a logical value.

Examples

# Create an empty graph; by default,
# new graphs made by `create_graph()`
# are directed
graph <-
  create_graph()

# Determine whether the graph
# is directed
graph %>%
  is_graph_directed()

# Use the `set_graph_undirected()`
# function and check again whether
# the graph is directed
graph %>%
  set_graph_undirected() %>%
  is_graph_directed()

[Package DiagrammeR version 1.0.1 Index]