is_graph_undirected {DiagrammeR}	R Documentation

Is the graph an undirected graph?

Description

Determines whether a graph is set as undirected or not and returns a logical value to that effect.

Usage

is_graph_undirected(graph)

Arguments

graph
a graph object of class dgr_graph.

Value

a logical value.

Examples

# Create an empty graph; by
# default, new graphs made
# by `create_graph()` are
# directed graph, so, use
# `directed = FALSE` to create
# an undirected graph
graph <-
  create_graph(
    directed = FALSE)

# Determine whether the
# graph is undirected
graph %>%
  is_graph_undirected()

# Use the `set_graph_directed()`
# function and check again
# as to whether the graph is
# undirected
graph %>%
  set_graph_directed() %>%
  is_graph_undirected()

[Package DiagrammeR version 1.0.1 Index]