node_data {DiagrammeR}	R Documentation

Insert node data attributes during node creation

Description

This helper function should be invoked to provide values for the namesake node_data argument, which is present in any function where nodes are created.

Usage

node_data(...)

Arguments

...
node data attributes provided as one or more named vectors.

Examples

# Create a new graph and add
# a path with several node
# data attributes
graph <-
  create_graph() %>%
  add_path(
    n = 3,
    type = "path",
    node_data = node_data(
      hour = 5,
      index = c(1, 3, 2)))

# View the graph's internal
# node data frame; the node
# data attributes have been
# inserted
graph %>%
  get_node_df()

[Package DiagrammeR version 1.0.1 Index]