rename_edge_attrs {DiagrammeR}	R Documentation

Rename an edge attribute

Description

Within a graph's internal edge data frame (edf), rename an existing edge attribute.

Usage

rename_edge_attrs(graph, edge_attr_from, edge_attr_to)

Arguments

graph
a graph object of class dgr_graph.

edge_attr_from
the name of the edge attribute that will be renamed.

edge_attr_to
the new name of the edge attribute column identified in edge_attr_from.

Value

a graph object of class dgr_graph.

Examples

# Create a random graph using the
# `add_gnm_graph()` function
graph <-
  create_graph() %>%
  add_gnm_graph(
    n = 5,
    m = 8,
    set_seed = 23) %>%
  set_edge_attrs(
    edge_attr = color,
    values = "green")

# Get the graph's internal edf
# to show which edge attributes
# are available
graph %>%
  get_edge_df()

# Rename the `color` node
# attribute as `weight`
graph <-
  graph %>%
  rename_edge_attrs(
    edge_attr_from = color,
    edge_attr_to = labelfontcolor)

# Get the graph's internal
# edf to show that the edge
# attribute had been renamed
graph %>%
  get_edge_df()

[Package DiagrammeR version 1.0.1 Index]