render_graph_from_graph_series {DiagrammeR}	R Documentation

Render a graph available in a series

Description

Using a graph series object of type dgr_graph_1D, either render graph in the Viewer or output in various formats.

Usage

render_graph_from_graph_series(graph_series, graph_no, output = "graph",
  width = NULL, height = NULL)

Arguments

graph_series
a graph series object of type dgr_graph_1D.

graph_no
the index of the graph in the graph series.

output
a string specifying the output type; graph (the default) renders the graph using the grViz function, DOT outputs DOT code for the graph, and SVG provides SVG code for the rendered graph.

width
an optional parameter for specifying the width of the resulting graphic in pixels.

height
an optional parameter for specifying the height of the resulting graphic in pixels.

Examples

## Not run: 
# Create three graphs
graph_1 <-
  create_graph() %>%
  add_path(n = 4)

graph_2 <-
  create_graph() %>%
  add_cycle(n = 5)

graph_3 <-
  create_graph() %>%
  add_star(n = 6)

# Create an empty graph series
# and add the graphs
series <-
  create_graph_series() %>%
  add_graph_to_graph_series(
    graph = graph_1) %>%
  add_graph_to_graph_series(
    graph = graph_2) %>%
  add_graph_to_graph_series(
    graph = graph_3)

# View the second graph in
# the series in the Viewer
render_graph_from_graph_series(
  graph_series = series,
  graph_no = 2)

## End(Not run)

[Package DiagrammeR version 1.0.1 Index]