select_nodes_by_id {DiagrammeR}	R Documentation

Select nodes in a graph by ID values

Description

Select nodes in a graph object of class dgr_graph by their node ID values. If nodes have IDs that are monotonically increasing integer values, then numeric ranges can be used for the selection.

Usage

select_nodes_by_id(graph, nodes, set_op = "union")

Arguments

graph
a graph object of class dgr_graph.

nodes
a vector of node IDs for the selection of nodes present in the graph.

set_op
the set operation to perform upon consecutive selections of graph nodes. This can either be as a union (the default), as an intersection of selections with intersect, or, as a difference on the previous selection, if it exists.

Value

a graph object of class dgr_graph.

Examples

# Create a node data frame (ndf)
ndf <-
  create_node_df(n = 10)

# Create a graph
graph <-
  create_graph(
    nodes_df = ndf)

# Select nodes `1` to `5` and show that
# selection of nodes with `get_selection()`
graph %>%
  select_nodes_by_id(nodes = 1:5) %>%
  get_selection()

[Package DiagrammeR version 1.0.1 Index]