select_nodes_in_neighborhood {DiagrammeR}	R Documentation

Select nodes based on a walk distance from a specified node

Description

Select those nodes in the neighborhood of nodes connected a specified distance from an initial node.

Usage

select_nodes_in_neighborhood(graph, node, distance, set_op = "union")

Arguments

graph
a graph object of class dgr_graph.

node
the node from which the traversal will originate.

distance
the maximum number of steps from the node for inclusion in the selection.

set_op
the set operation to perform upon consecutive selections of graph nodes. This can either be as a union (the default), as an intersection of selections with intersect, or, as a difference on the previous selection, if it exists.

Value

a graph object of class dgr_graph.

Examples

# Create a graph containing
# a balanced tree
graph <-
  create_graph() %>%
  add_balanced_tree(
    k = 2, h = 2)

# Create a graph selection by
# selecting nodes in the
# neighborhood of node `1`, where
# the neighborhood is limited by
# nodes that are 1 connection
# away from node `1`
graph <-
  graph %>%
  select_nodes_in_neighborhood(
    node = 1,
    distance = 1)

# Get the selection of nodes
graph %>%
  get_selection()

# Perform another selection
# of nodes, this time with a
# neighborhood spanning 2 nodes
# from node `1`
graph <-
  graph %>%
  clear_selection() %>%
  select_nodes_in_neighborhood(
    node = 1,
    distance = 2)

# Get the selection of nodes
graph %>%
  get_selection()

[Package DiagrammeR version 1.0.1 Index]