set_cache {DiagrammeR}	R Documentation

Cache a vector in the graph

Description

Place any vector in the cache of a graph object of class dgr_graph.

Usage

set_cache(graph, to_cache, name = NULL, col = NULL)

Arguments

graph
a graph object of class dgr_graph.

to_cache
any vector or data frame. If a data frame is supplied then a single column for the vector to pull must be provided in the col argument.

name
an optional name for the cached vector.

col
if a data frame is provided in to_cache then a column name from that data frame must provided here.

Value

a graph object of class dgr_graph.

Examples

# Create a random graph using the
# `add_gnm_graph()` function
graph <-
  create_graph() %>%
  add_gnm_graph(
    n = 10,
    m = 22,
    set_seed = 23)

# Get the closeness values for
# all nodes from `1` to `10` and
# store in the graph's cache
graph <-
  graph %>%
  set_cache(
    name = "closeness_vector",
    to_cache = get_closeness(.),
    col = "closeness")

# Get the graph's cache
graph %>%
  get_cache(
    name = "closeness_vector")

# Get the difference of betweenness
# and closeness values for nodes in
# the graph and store the vector in
# the graph's cache
graph <-
  graph %>%
  set_cache(
    name = "difference",
    to_cache =
      get_betweenness(.)$betweenness -
        get_closeness(.)$closeness)

# Get the graph's cache
graph %>%
  get_cache(
    name = "difference")

[Package DiagrammeR version 1.0.1 Index]