transform_to_complement_graph {DiagrammeR}	R Documentation

Create a complement of a graph

Description

Create a complement graph which contains only edges not present in the input graph. It's important to nodes that any edge attributes in the input graph's edges will be lost. Node attributes will be retained, since they are not affected by this transformation.

Usage

transform_to_complement_graph(graph, loops = FALSE)

Arguments

graph
a graph object of class dgr_graph that is created using create_graph.

loops
an option for whether loops should be generated in the complement graph.

Value

a graph object of class dgr_graph.

Examples

# Create a simple graph
# with a single cycle
graph <-
  create_graph() %>%
  add_cycle(n = 4)

# Get the graph's edge
# data frame
graph %>%
  get_edge_df()

# Create the complement
# of the graph
graph_c <-
  graph %>%
    transform_to_complement_graph()

# Get the edge data frame
# for the complement graph
graph_c %>%
  get_edge_df()

[Package DiagrammeR version 1.0.1 Index]