usd_exchange_rates {DiagrammeR}	R Documentation

US Dollar exchange rates.

Description

A dataset containing exchange rates from USD to all other currencies.

Usage

usd_exchange_rates

Format

A data frame with 196 rows and 3 variables:

from_currency

the currency from which units will be used to buy units in the alternate currency (this is always USD)

to_currency

the currency that is to be bought

cost_unit

the cost per unit of the currency to be bought

[Package DiagrammeR version 1.0.1 Index]