count_automorphisms {DiagrammeR}	R Documentation

Get the number of automorphisms

Description

Get the number of automorphisms the graph contains. An automorphism of a graph is a form of symmetry in which the graph is mapped onto itself while preserving edge-node connectivity.

Usage

count_automorphisms(graph)

Arguments

graph
a graph object of class dgr_graph.

Value

a single numeric value representing the number of automorphisms the graph contains.

Examples

# Create a cycle graph
graph <-
  create_graph() %>%
  add_cycle(n = 5)

# Get a count of automorphisms
graph %>%
  count_automorphisms()

# Create a full graph and then
# count the automorphisms
create_graph() %>%
  add_full_graph(n = 10) %>%
  count_automorphisms()

[Package DiagrammeR version 1.0.1 Index]