Delete a node from an existing graph object

From a graph object of class dgr_graph, delete an existing node by specifying its node ID.

Usage

delete_node(graph, node)

Arguments

graph

A graph object of class dgr_graph.

node

A node ID for the node to be deleted from the graph.

Value

A graph object of class dgr_graph.

See also

Examples

# Create a graph with 5 nodes and
# edges between each in a path
graph <-
  create_graph() %>%
  add_path(n = 5)

# Delete node with ID `3`
graph <- delete_node(graph, node = 3)

# Verify that the node with ID `3`
# is no longer in the graph
graph %>% get_node_ids()
#> [1] 1 2 4 5

# Also note that edges are removed
# since there were edges between the
# removed node to and from other nodes
graph %>% get_edges()
#> [1] "1->2" "4->5"