from_igraph {DiagrammeR}	R Documentation

Convert an igraph graph to a DiagrammeR one

Description

Convert an igraph graph to a DiagrammeR graph object.

Usage

from_igraph(igraph, graph_name = NULL, write_backups = FALSE)

Arguments

igraph
an igraph graph object.

graph_name
an optional string for labeling the graph object.

write_backups
an option to write incremental backups of changing graph states to disk. If TRUE, a subdirectory of the working directory will be used to store RDS files. The default value is FALSE so one has to opt in to use this functionality.

Value

a graph object of class dgr_graph.

Examples

# Create a DiagrammeR graph object
dgr_graph_orig <-
  create_graph() %>%
  add_gnm_graph(
    n = 36,
    m = 50,
    set_seed = 23)

# Convert the DiagrammeR
# graph to an igraph object
ig_graph <-
  dgr_graph_orig %>%
  to_igraph()

# Convert the igraph graph
# back to a DiagrammeR graph
dgr_graph_new <-
  ig_graph %>%
  from_igraph()

# Get some graph information
(dgr_graph_new %>%
  get_graph_info())[, 1:6]

[Package DiagrammeR version 1.0.1 Index]