get_edge_info {DiagrammeR}	R Documentation

Get detailed information on edges

Description

Obtain a data frame with detailed information on edges and their interrelationships within the graph.

Usage

get_edge_info(graph)

Arguments

graph
a graph object of class dgr_graph.

Value

a data frame containing information specific to each edge within the graph.

Examples

# Create a simple graph
graph <-
  create_graph() %>%
  add_gnm_graph(
    n = 5, m = 10,
    set_seed = 23)

# Get information on the
# graph's edges
graph %>%
  get_edge_info()

[Package DiagrammeR version 1.0.1 Index]