get_girth {DiagrammeR}	R Documentation

Get graph girth

Description

Get the girth of a graph, which is the length of the shortest circle in the graph. Loop edges and multiple edges are not considered. If the graph contains no cycles then zero is returned.

Usage

get_girth(graph)

Arguments

graph
a graph object of class dgr_graph.

Value

a single numeric value representing the length of the shortest circle in the graph.

Examples

# Create a cycle graph
graph <-
  create_graph() %>%
  add_cycle(n = 5)

# Determine the graph's girth
graph %>%
  get_girth()

# Create a full graph and then
# get the girth for that
create_graph() %>%
  add_full_graph(n = 10) %>%
  get_girth()

[Package DiagrammeR version 1.0.1 Index]