get_graph_log {DiagrammeR}	R Documentation

Get the graph log information

Description

Get a tibble of the graph log, which contains information on the functions called on the graph that resulted in some transformation of the graph.

Usage

get_graph_log(graph)

Arguments

graph
a graph object of class dgr_graph.

Value

a df_tbl object.

Examples

# Create a random graph using the
# `add_gnm_graph()` function and
# delete 2 nodes from the graph
graph <-
  create_graph(
    directed = FALSE) %>%
  add_gnm_graph(
    n = 10,
    m = 15,
    set_seed = 23) %>%
  delete_node(node = 5) %>%
  delete_node(node = 7)

# Get the graph log, which is a
# record of all graph transformations
graph %>%
  get_graph_log()

[Package DiagrammeR version 1.0.1 Index]