Get the graph log information

Get a tibble of the graph log, which contains information on the functions called on the graph that resulted in some transformation of the graph.

get_graph_log(graph)

Arguments

graph
a graph object of class dgr_graph.

Value

a df_tbl object.

Examples

# Create a random graph using the
# `add_gnm_graph()` function and
# delete 2 nodes from the graph
graph <-
  create_graph(
    directed = FALSE) %>%
  add_gnm_graph(
    n = 10,
    m = 15,
    set_seed = 23) %>%
  delete_node(node = 5) %>%
  delete_node(node = 7)

# Get the graph log, which is a
# record of all graph transformations
graph %>%
  get_graph_log()

#> # A tibble: 4 x 8
#>   version_id function_used       time_modified    duration nodes edges   d_n
#>        <int>         <chr>              <dttm>       <dbl> <int> <int> <int>
#> 1          1  create_graph 2018-01-25 01:36:59 0.012403965     0     0     0
#> 2          2 add_gnm_graph 2018-01-25 01:36:59 0.016880989    10    15    10
#> 3          3   delete_node 2018-01-25 01:36:59 0.008655071     9    11    -1
#> 4          4   delete_node 2018-01-25 01:36:59 0.008348942     8     8    -1
#> # ... with 1 more variables: d_e <int>