Get node attribute values from a selection of nodes

From a graph object of class dgr_graph, get node attribute values from nodes currently active as a selection.

get_node_attrs_ws(graph, node_attr)

Arguments

graph
a graph object of class dgr_graph.

node_attr
the name of the attribute for which to get values.

Value

a named vector of node attribute values for the attribute given by node_attr by node ID.

Details

This function makes use of an active selection of nodes (and the function ending with _ws hints at this).

Selections of nodes can be performed using the following node selection (select_*()) functions: select_nodes(), select_last_nodes_created(), select_nodes_by_degree(), select_nodes_by_id(), or select_nodes_in_neighborhood().

Selections of nodes can also be performed using the following traversal (trav_*()) functions: trav_out(), trav_in(), trav_both(), trav_out_node(), trav_in_node(), trav_out_until(), or trav_in_until().

Examples

# Create a random graph using the
# `add_gnm_graph()` function
graph <-
  create_graph() %>%
  add_gnm_graph(
    n = 4,
    m = 4,
    set_seed = 23) %>%
  set_node_attrs(
    node_attr = value,
    values = c(2.5, 8.2, 4.2, 2.4))

# Select nodes with ID values
# `1` and `3`
graph <-
  graph %>%
  select_nodes_by_id(
    nodes = c(1, 3))

#> `select_nodes_by_id()` INFO: created a new selection of 2 nodes

# Get the node attribute values
# for the `value` attribute, limited
# to the current node selection
graph %>%
  get_node_attrs_ws(
    node_attr = value)

#>   1   3 
#> 2.5 4.2