Get the current selection available in a graph object

Get the current selection of node IDs or edge IDs from a graph object of class dgr_graph.

get_selection(graph)

Arguments

graph
a graph object of class dgr_graph.

Value

a vector with the current selection of node or edge ID values.

Examples

# Create a simple graph
graph <-
  create_graph() %>%
  add_path(n = 6)

# Select node `4`, then select
# all nodes a distance of 1 away
# from node `4`, and finally
# return the selection of nodes as
# a vector object
graph %>%
  select_nodes(nodes = 4) %>%
  select_nodes_in_neighborhood(
    node = 4,
    distance = 1) %>%
  get_selection()

#> `select_nodes()` INFO: created a new selection of 1 node

#> `select_nodes_in_neighborhood()` INFO: created a new selection of3 nodes:
#> * this replaces1 nodein the prior selection

#> `get_selection()` INFO: there is an active selection of 3 nodes

#> [1] 3 4 5

# Select edges associated with
# node `4` and return the
# selection of edges
graph %>%
  select_edges_by_node_id(
    nodes = 4) %>%
  get_selection()

#> `select_edges_by_node_id()` INFO: created a new selection of 2 edges

#> `get_selection()` INFO: there is an active selection of2 edges

#> [1] 3 4