get_w_connected_cmpts {DiagrammeR}	R Documentation

Get all nodes associated with connected components

Description

Determine which nodes in a graph belong to different weakly connected components (i.e., distinct sets of nodes with traversable paths to and from each node in the set).

Usage

get_w_connected_cmpts(graph)

Arguments

graph
a graph object of class dgr_graph.

Value

a data frame with nodes and their membership in different weakly connected components.

Examples

# Create a graph with 2 cycles
graph <-
  create_graph() %>%
  add_cycle(n = 4) %>%
  add_cycle(n = 3)

# Check if the graph is connected
graph %>%
  is_graph_connected()

# Get the graph's weakly-connected
# components
graph %>%
  get_w_connected_cmpts()

[Package DiagrammeR version 1.0.1 Index]