is_graph_weighted {DiagrammeR}	R Documentation

Is the graph a weighted graph?

Description

Provides a logical value on whether the graph is weighted. A graph is considered to be weighted when it contains edges that all have a edge weight attribute with numerical values assigned for all edges.

Usage

is_graph_weighted(graph)

Arguments

graph
a graph object of class dgr_graph.

Value

a logical value.

Examples

# Create a graph where the edges have
# a `weight` attribute
graph <-
  create_graph() %>%
  add_cycle(n = 5) %>%
  select_edges() %>%
  set_edge_attrs_ws(
    edge_attr = weight,
    value = c(3, 5, 2, 9, 6)) %>%
  clear_selection()

# Determine whether the graph
# is a weighted graph
graph %>%
  is_graph_weighted()

# Create graph where the edges do
# not have a `weight` attribute
graph <-
  create_graph() %>%
  add_cycle(n = 5)

# Determine whether this graph
# is weighted
graph %>%
  is_graph_weighted()

[Package DiagrammeR version 1.0.1 Index]