Reverse the direction of all edges in a graph

Using a directed graph as input, reverse the direction of all edges in that graph.

Usage

rev_edge_dir(graph)

Arguments

graph

A graph object of class dgr_graph.

Value

A graph object of class dgr_graph.

See also

Examples

# Create a graph with a
# directed tree
graph <-
  create_graph() %>%
  add_balanced_tree(
    k = 2, h = 2)

# Inspect the graph's edges
graph %>% get_edges()
#> [1] "1->2" "1->3" "2->4" "2->5" "3->6" "3->7"

# Reverse the edge directions
# such that edges are directed
# toward the root of the tree
graph <-
  graph %>%
  rev_edge_dir()

# Inspect the graph's edges
# after their reversal
graph %>% get_edges()
#> [1] "2->1" "3->1" "4->2" "5->2" "6->3" "7->3"