select_last_edges_created {DiagrammeR}	R Documentation

Select the last set of edges created in a graph

Description

Select the last edges that were created in a graph object of class dgr_graph. This function should ideally be used just after creating the edges to be selected.

Usage

select_last_edges_created(graph)

Arguments

graph
a graph object of class dgr_graph.

Value

a graph object of class dgr_graph.

Examples

# Create a graph and add a cycle and then
# a tree in 2 separate function calls
graph <-
  create_graph() %>%
  add_cycle(
    n = 3,
    rel = "a") %>%
  add_balanced_tree(
    k = 2, h = 2,
    rel = "b")

# Select the last edges created (all edges
# from the tree) and then set their edge
# color to be `red`
graph <-
  graph %>%
  select_last_edges_created() %>%
  set_edge_attrs_ws(
    edge_attr = color,
    value = "red") %>%
  clear_selection()

# Display the graph's internal edge
# data frame to verify the change
graph %>%
  get_edge_df()

[Package DiagrammeR version 1.0.1 Index]