set_edge_attrs_ws {DiagrammeR}	R Documentation

Set edge attributes with an edge selection

Description

From a graph object of class dgr_graph or an edge data frame, set edge attribute properties for one or more edges.

Usage

set_edge_attrs_ws(graph, edge_attr, value)

Arguments

graph
A graph object of class dgr_graph.

edge_attr
The name of the attribute to set.

value
The value to be set for the chosen attribute for the edges in the current selection.

Details

This function makes use of an active selection of edges (and the function ending with _ws hints at this).

Selections of edges can be performed using the following selection (select_*()) functions: select_edges(), select_last_edges_created(), select_edges_by_edge_id(), or select_edges_by_node_id().

Selections of edges can also be performed using the following traversal (trav_*()) functions: trav_out_edge(), trav_in_edge(), trav_both_edge(), or trav_reverse_edge().

Value

A graph object of class dgr_graph.

Examples

# Create a simple graph
graph <-
  create_graph() %>%
  add_path(n = 6)

# Select specific edges from
# the graph and apply the edge
# attribute `color = blue` to
# those selected edges
graph <-
  graph %>%
  select_nodes_by_id(nodes = 2:4) %>%
  trav_out_edge() %>%
  set_edge_attrs_ws(
    edge_attr = color,
    value = "blue")

# Show the internal edge data
# frame to verify that the
# edge attribute has been set
# for specific edges
graph %>% get_edge_df()

[Package DiagrammeR version 1.0.5 Index]