Set node attributes with a node selection

From a graph object of class dgr_graph or a node data frame, set node attribute properties for nodes present in a node selection.

Selections of nodes can be performed using the following select_... functions: select_nodes(), select_last_nodes_created(), select_nodes_by_degree(), select_nodes_by_id(), or select_nodes_in_neighborhood(). Selections of nodes can also be performed using the following traversal functions: (trav_...): trav_out(), trav_in(), trav_both(), trav_in_node(), trav_out_node().

set_node_attrs_ws(graph, node_attr, value)

Arguments

graph
a graph object of class dgr_graph.

node_attr
the name of the attribute to set.

value
the value to be set for the chosen attribute for the nodes in the current selection.

Value

a graph object of class dgr_graph.

Examples

# Create a simple graph
graph <-
  create_graph() %>%
  add_path(n = 6)

# Select specific nodes from the graph and
# apply the node attribute `color = blue` to
# those selected nodes
graph <-
  graph %>%
  select_nodes_by_id(
    nodes = 1:4) %>%
  trav_out() %>%
  set_node_attrs_ws(
    node_attr = color,
    value = "blue")

# Show the internal node data frame to verify
# that the node attribute has been set for
# specific node
graph %>%
  get_node_df()

#>   id type label color
#> 1  1 <NA>     1  <NA>
#> 2  2 <NA>     2  blue
#> 3  3 <NA>     3  blue
#> 4  4 <NA>     4  blue
#> 5  5 <NA>     5  blue
#> 6  6 <NA>     6  <NA>